requirements/common.txt
asgiref==3.7.2
bcrypt==4.1.2
boto3==1.34.30
botocore==1.34.30
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.1
django-bootstrap-icons==0.8.7
django-bootstrap5==23.4
django-filter==23.5
django-prettyjson==0.4.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
jmespath==1.0.1
Markdown==3.5.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
standardjson==0.3.1
urllib3==2.0.7
gunicorn==21.2.0
